testing of background tile rendering

+++ Farm Sim/src/specifier/Vector2D.java
@@ -2,9 +2,9 @@ package specifier;
 
 public class Vector2D
 {
-	public float x, y;
+	public int x, y;
 	
-	public Vector2D(float x, float y)
+	public Vector2D(int x, int y)
 	{
 		this.x = x;
 		this.y = y;
